[HOPS-860] Parallelized -copyToLocal HDFS CLI
