+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/CommonConfigurationKeys.java
@@ -380,7 +380,7 @@ public class CommonConfigurationKeys extends CommonConfigurationKeysPublic {
   public static final String IO_ERASURECODE_CODEC_XOR_RAWCODER_DEFAULT =
       XORRawErasureCoderFactory.class.getCanonicalName();
 
-  public static final String DFS_CLIENT_COPY_FROM_LOCAL_PARALLEL_THREADS =
-          "dfs.client.copy.from.local.parallel.threads";
-  public static final int DFS_CLIENT_COPY_FROM_LOCAL_PARALLEL_THREADS_DEFAULT = 10;
+  public static final String DFS_CLIENT_COPY_TO_OR_FROM_LOCAL_PARALLEL_THREADS =
+          "dfs.client.copy.to.or.from.local.parallel.threads";
+  public static final int DFS_CLIENT_COPY_TO_OR_FROM_LOCAL_PARALLEL_THREADS_DEFAULT = 10;
 }
